# [PERF] pos_loyalty: improve perf of _compute_pos_order_count