Current Behavior:

Computation of the field `pos_order_count` on model `loyalty.program` is slow when there is a large number of `pos.order.line` records with `reward_id` set on them.

Cause of the issue:

The list comprehension done here --> https://github.com/odoo/odoo/blob/27ff3e0f64f53caa62c3022bd3b7c41e29a8e721/addons/pos_loyalty/models/loyalty_program.py#L59 The complexity is `O(len(self) * ((len(read_group_res) * len(program_reward_ids)) + len(read_group_res<sum method>)))` which performs slowly if the `self` and `read_group_res`  are large.

Improvement:

Delegate the computation to Postgres and assign the values obtained from the result.

Benchmark:

In method `_compute_pos_order_count`
Number of `pos.order.line` records eligible in computation -->  17803
Where `self` is a a `loyalty.program` recordset

len(self)| Before (in seconds) | After PR (in seconds) |
 |---------|--------|--------|
|1000| 177.28 s | 6.43 s |
|212| 29.73 s| 0.5 s |

Improvement by about 98% on average

opw-3903159