@adda-odoo 2024-07-01
Hi @caburj! Any chance to have this PR reviewed soon? Thanks!